refactor(#255): one export per file in service-test-utils bun harness
Splits every multi-export module (create-database-from-template,
env-overrides, the isolation strategies) into one file per export, renames
TestDB to TestDBHandle so its filename matches, and drops comments that
named other declarations by exact identifier.
