+++ projects/lib-service-test-utils/src/bun/create-database-from-template.ts
@@ -1,26 +1,6 @@
 import { createId } from '@paralleldrive/cuid2';
 import postgres from 'postgres';
-
-const DEFAULT_TEST_DB_URI = 'postgres://test:test@localhost:32999';
-const DEFAULT_TEST_TEMPLATE_DB = 'test_template';
-
-/** The test container's base connection URI and migrated template database name. */
-interface TestDBTarget {
-  readonly baseURI: string;
-  readonly templateDB: string;
-}
-
-/**
- * Resolves the test container's base URI and migrated template DB name, published by
- * `setupBunTestDB` as `TEST_DB_URI`/`TEST_TEMPLATE_DB`. Falls back to the test container's fixed
- * defaults so a bare `bun test` against an already-running container still resolves.
- */
-export function resolveTestDBTarget(): TestDBTarget {
-  return {
-    baseURI: process.env['TEST_DB_URI'] ?? DEFAULT_TEST_DB_URI,
-    templateDB: process.env['TEST_TEMPLATE_DB'] ?? DEFAULT_TEST_TEMPLATE_DB,
-  };
-}
+import { resolveTestDBTarget } from './resolve-test-db-target';
 
 /** Creates a new database cloned from the migrated template; returns its connection URL. */
 export async function createDatabaseFromTemplate(): Promise<string> {
