+++ projects/lib-service-test-utils/src/bun/original-env.ts
@@ -1,0 +1,2 @@
+/** Each overridden key's value from before its first override, `undefined` when it was unset. */
+export const originals = new Map<string, string | undefined>();
