+++ projects/lib-service-test-utils/src/bun/strategies/create-transaction-test-db.ts
@@ -2,7 +2,7 @@ import { createDB } from '@vers/db';
 import type { DB } from '@vers/db';
 import type { Kysely } from 'kysely';
 import { createDatabaseFromTemplate } from '../create-database-from-template';
-import type { TestDB } from '../test-db-handle';
+import type { TestDBHandle } from '../test-db-handle';
 
 /**
  * Transaction isolation: rollback on dispose, near-zero per-test cost. Correct only for code that
@@ -11,7 +11,7 @@ import type { TestDB } from '../test-db-handle';
  * CONFLICT/unique error must not issue further queries on the same handle; it asserts the error
  * and ends.
  */
-export async function createTransactionTestDB(): Promise<TestDB> {
+export async function createTransactionTestDB(): Promise<TestDBHandle> {
   const db = await getWorkerDB();
   const trx = await db.startTransaction().execute();
